+++ table/block_based/parsed_full_filter_block.cc
@@ -5,7 +5,8 @@
 //
 
 #include "table/block_based/parsed_full_filter_block.h"
-#include "rocksdb/filter_policy.h"
+
+#include "table/block_based/filter_policy_internal.h"
 
 namespace ROCKSDB_NAMESPACE {
